src/asr/requirements.txt
torchvision
scikit-image
pyannote.algorithms
sidekit
librosa
mpi4py
deepspeed